+++ VersionHistory.md
@@ -8,6 +8,11 @@ Prefix the description of the change with `[major]`, `[minor]`, or `[patch]` in 
 
 ## Released
 
+### 3.8.0
+
+* [minor] Support C# 8 nullable references.
+* [minor] Support .NET local tools.
+
 ### 3.7.4
 
 * [patch] Fix Visual Studio 2019 path.